refactor: context 구조 정리
